Реализовано удаление выбранной услуги.

+++ app/src/main/java/com/cespaul/arent/model/RentService.kt
@@ -3,7 +3,7 @@ package com.cespaul.arent.model
 data class RentService(
     var id: Int,
     var nameService: String,
-    var rateService: Int,
-    var amtService: Int,
-    var sumService: Int
+    var rateService: Float,
+    var amtService: Float,
+    var sumService: Float
 )
